Add files via upload
The comprehensive analysis using SQL and the creation of dynamic executive dashboards in Excel and Tableau provide valuable insights into the bike store's operations. These tools facilitate data-driven decision-making, helping the store optimize sales strategies, improve customer engagement, and manage inventory effectively. The visualizations and interactive elements in the dashboards ensure that complex data is presented in an accessible and actionable format for the management team.